+++ geograpix/src/App.js
@@ -4,12 +4,14 @@ import Map from './components/Map';
 
 import Register from './components/Register/Register';
 import Register2 from './components/Register/Register2';
+import Preloader from './components/Preloader';
 
 function App() {
 
   return (
     <div className="App">
       <Route exact path="/" component={Register} />
+      <Route exact path="/preloader" render={(props) => <Preloader {...props} />} />
       <Route path="/register/2/" component={Register2} />
       <section class="map">
         <header class="App-header">
